interface audio: changed ints to uint8_t

+++ interface_alsa.h
@@ -30,22 +30,22 @@ public:
     
     void createNewPort(QString label);
     
-    void keyPressEvent(int port, int channel, int midicode, int velocity);
-    void keyReleaseEvent(int port, int channel, int midicode, int velocity);
+    void keyPressEvent(int port, int channel, uint8_t midicode, uint8_t velocity);
+    void keyReleaseEvent(int port, int channel, uint8_t midicode, uint8_t velocity);
     void keyPanicEvent(int port, int channel);
     void keyStopAllEvent(int port, int channel);
     void keyPitchbendEvent(int port, int channel, int pitch);
     void keySustainEvent(int port, int channel, bool pressed);
     void keySostenutoEvent(int port, int channel, bool pressed);
     void keySoftEvent(int port, int channel, bool pressed);
-    void setProgramChangeEvent(int port, int channel, int program, int bank);
-    void setControlChangeEvent(int port, int channel, int cc, int value);
-    void setVolumeChangeEvent(int port, int channel, int volume);
-    void setPanChangeEvent(int port, int channel, int value);
-    void setPortamentoChanged(int port, int channel, int value);
-    void setAttackChanged(int port, int channel, int value);
-    void setReleaseChanged(int port, int channel, int value);
-    void setTremoloChanged(int port, int channel, int value);
+    void setProgramChangeEvent(int port, int channel, uint8_t program, int bank);
+    void setControlChangeEvent(int port, int channel, uint8_t cc, uint8_t value);
+    void setVolumeChangeEvent(int port, int channel, uint8_t volume);
+    void setPanChangeEvent(int port, int channel, uint8_t value);
+    void setPortamentoChanged(int port, int channel, uint8_t value);
+    void setAttackChanged(int port, int channel, uint8_t value);
+    void setReleaseChanged(int port, int channel, uint8_t value);
+    void setTremoloChanged(int port, int channel, uint8_t value);
     
     QString NAME = "vkeybd-qt";
